extract help command
